+++ src/types.ts
@@ -1,0 +1,7 @@
+export type Message = {
+  from: { email: string; name: string }
+  to: { email: string; name: string }[]
+  data?: { [key: string]: any }[]
+  subject: string
+  content: string
+}